move getWorkspaceUtilities
